pyproject.toml
[project]
name = "EduPress"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "celery>=5.5.3",
    "django>=5.2.6",
    "django-celery-results>=2.6.0",
    "django-ckeditor-5>=0.2.18",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "django-location-field>=2.7.3",
    "django-minio-backend>=3.8.1",
    "django-nested-admin>=4.1.5",
    "django-video-encoding>=1.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "faker>=37.6.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "six>=1.17.0",
]
